Move classes into config package
